requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML dependencies
scikit-learn>=1.3.0
pandas==2.1.4
joblib>=1.3.0

# Data processing
requests>=2.28.0

# MLflow для версионирования моделей
mlflow==2.8.1

# Airflow для автоматизации pipeline
apache-airflow==2.7.3

# Spark для обработки больших данных
pyspark==3.5.0

# PyTorch для глубокого обучения
torch==2.1.1
torchtext==0.16.1
transformers==4.35.2

# Дополнительные библиотеки для работы с данными
pyarrow==14.0.1
findspark==2.0.1
mlflow>=2.8.0

# Optional: Kaggle for data downloads
kaggle>=1.5.12